+++ src/algitex/algo/loop.py
@@ -1,4 +1,4 @@
 """Re-export Loop."""
-from devloop.algo import Loop
+from algitex.algo import Loop
 
 __all__ = ["Loop"]