+++ src/algitex/propact/workflow.py
@@ -1,0 +1,2 @@
+from algitex.propact import Workflow
+__all__ = ["Workflow"]